Commit changes moving useUser auth stuff to useAuth
