+++ apps/web/src/pages/overview/Overview.vue
@@ -1,20 +1,19 @@
 <script setup lang="ts">
 import { ref, watch } from 'vue'
 import Staking from './components/Staking.vue'
-import Announcements from './components/Announcements.vue'
 import StakingAvg from './components/StakingAvg.vue'
 import BreakdownChart from './components/BreakdownChart.vue'
 import BreakdownTable from './components/BreakdownTable.vue'
 
-import useUser from '@/composables/user'
+import useAuth from '@/composables/auth'
 import useBreakdownMetrics from '@/composables/breakdownMetrics'
 import useAnalytics from '@/composables/analytics'
 import useOperators from '@/composables/operators'
 
-const {loadingSessionLogin} = useUser()
-const {loadingInitializeBreakdownMetrics} = useBreakdownMetrics()
-const {loadingInitializeAnalytics} = useAnalytics()
-const {loadingInitializeOperators} = useOperators()
+const { loadingSessionLogin } = useAuth()
+const { loadingInitializeBreakdownMetrics } = useBreakdownMetrics()
+const { loadingInitializeAnalytics } = useAnalytics()
+const { loadingInitializeOperators } = useOperators()
 
 const showSkeleton = ref(true)
